+++ API/app/Http/Controllers/Api/CardController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Models\Card;
 use App\Models\Log;
+use App\Models\Order;
+use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Http;
@@ -215,6 +217,10 @@ class CardController extends Controller
         $user_authenticaded = JWTAuth::toUser();
 
         try{
+            $order = Order::where('card_id', $id)->first();
+            if(!is_null($order)){
+                throw new Exception();
+            }
             $result = $user_authenticaded->card()->where('id', $id)->delete();
         }catch(\Exception $exception){
             $error = ['code' => 2, 'error_message' => 'Não foi possivel deletar o cartão.'];
